+++ packages/components/src/Container/Container.tsx
@@ -1,10 +1,9 @@
 import React from 'react';
+import { ComponentProps } from '@marigold/types';
+
 import { Box } from '../Box';
 
-export type ContainerProps = {
-  className?: string;
-  title?: string; // Used for testing.
-};
+export interface ContainerProps extends ComponentProps<'div'> {}
 
 export const Container: React.FC<ContainerProps> = ({ children, ...props }) => (
   <Box {...props} width="100%">