+++ packages/components/src/Stack/Stack.tsx
@@ -1,14 +1,13 @@
 import React, { Children } from 'react';
-import flattenChildren from 'react-keyed-flatten-children';
-
 import { ResponsiveStyleValue } from '@marigold/system';
 
+import { flattenChildren } from '../utils';
 import { Box } from '../Box';
 
-export type StackProps = {
+export interface StackProps {
   space?: ResponsiveStyleValue<string>;
   align?: 'left' | 'right' | 'center';
-};
+}
 
 const ALIGNMENT = {
   left: 'flex-start',